Allow multiple instances when running ui-tests
